CA2: Migrated dpenalty from float to int. Optimize SAS correction by not evaluating melodies, that already have dpenalty larger than dpenalty_min (closed #1079)

+++ MGen/Generators/GenCA2.cpp
@@ -46,7 +46,7 @@ void CGenCA2::SendCorrectionsCP(int i, long long time_start) {
 	for (int p = 0; p < corrections; p++) {
 		// Find minimum penalty
 		cids.clear();
-		float dpenalty_min = MAX_PENALTY;
+		int dpenalty_min = MAX_PENALTY;
 		for (int x = 0; x < cnum; x++) if (dpenalty[x] < dpenalty_min) dpenalty_min = dpenalty[x];
 		// Get all best corrections
 		for (int x = 0; x < cnum; x++) if (dpenalty[x] == dpenalty_min) {
@@ -60,7 +60,7 @@ void CGenCA2::SendCorrectionsCP(int i, long long time_start) {
 			ccount++;
 			if (ccount > corrections) break;
 			// Write log
-			st.Format("%.0f/%.0f/%d/%d ", rpenalty_min, dpenalty_min, cids.size(), clib.size());
+			st.Format("%.0f/%d/%d/%d ", rpenalty_min, dpenalty_min, cids.size(), clib.size());
 			st2 += st;
 			// Show initial melody again if this is not first iteration
 			if (ccount > 1) {
